pyproject.toml
[project]
name = "text-to-sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "matplotlib>=3.10.8",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
